my-app/backend/requirements.txt
absl-py==2.1.0
attrs==23.2.0
basemap==1.4.1
basemap-data==1.3.2
basemap-data-hires==1.3.2
blinker==1.8.2
bokeh==3.5.2
cachelib==0.9.0
cachetools==5.5.0
Cartopy==0.23.0
certifi==2024.8.30
cffi==1.16.0
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorcet==3.1.0
contourpy==1.2.1
cycler==0.12.1
dask==2024.9.0
dask-expr==1.1.14
datashader==0.16.3
dnspython==2.6.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.9.0
geopandas==1.0.1
gevent==24.2.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h5netcdf==1.3.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
jax==0.4.26
jaxlib==0.4.26
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mediapipe==0.10.13
ml-dtypes==0.4.0
multipledispatch==1.0.0
netCDF4==1.7.1.post2
numba==0.60.0
numpy==2.1.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
param==2.1.1
partd==1.4.2
pillow==10.3.0
plotly==5.24.1
protobuf==4.25.3
pyarrow==17.0.0
pycparser==2.22
pyct==0.5.0
Pygments==2.18.0
pymongo==4.10.1
pyngrok==7.2.0
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
requests==2.32.3
rich==13.8.1
scipy==1.13.0
seaborn==0.13.2
selenium==4.25.0
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
tenacity==9.0.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
xarray==2024.9.0
xyzservices==2024.9.0
zope.event==5.0
zope.interface==7.0.3
python-dotenv
